KE-9016 convert double to long

+++ core/src/main/java/org/apache/calcite/runtime/SqlFunctions.java
@@ -1730,7 +1730,7 @@ public class SqlFunctions {
     if (s.startsWith("199") && s.contains(":")) {
       return Timestamp.valueOf(s).getTime();
     }
-    return Long.parseLong(s.trim());
+    return (long) Double.parseDouble(s.trim());
   }
 
   public static long toLong(Number number) {
